Перенести метод loadFromFile в класс FileBackedTaskManager
