+++ tasks.csv
@@ -1,5 +1,4 @@
 id,type,name,status,description,epic or subtasks
-1,TASK,Задача 1,NEW,Описание задачи 1,
 3,SUBTASK,Подзадача 1,NEW,Описание подзадачи 1,2
 4,SUBTASK,Подзадача 2,NEW,Описание подзадачи 2,2
 2,EPIC,Эпик 1,NEW,Описание Эпика 1,3 4
