+++ test.csv
@@ -4,4 +4,4 @@ id,type,name,status,description,epic or subtasks
 4,SUBTASK,Подзадача 2,NEW,Описание подзадачи 2,2
 2,EPIC,Эпик 1,NEW,Описание Эпика 1,
 
-1,3
+empty history
